fix: recursive option not work when directly call versionBump
